requirements.txt
asgiref==3.2.10
certifi==2020.6.20
click==7.1.2
Django==3.1.1
django-environ==0.4.5
djangorestframework==3.12.1
flake8==3.8.3
greenlet==0.4.17
gunicorn==20.0.4
h11==0.10.0
httptools==0.1.1
mccabe==0.6.1
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
structlog==20.1.0
uvicorn==0.12.0
uvloop==0.14.0
